requirements.txt
numpy>=1.24
pillow>=10.0
opencv-python>=4.8
torch>=2.3
torchvision>=0.18
transformers>=4.41
tqdm>=4.66
requests>=2.31
